更改 guide模式 电子围栏判断

+++ libraries/AP_Math/location.cpp
@@ -214,11 +214,11 @@ void print_latlon(AP_HAL::BetterStream *s, int32_t lat_or_lon)
 // return true when lat and lng are within range
 bool check_lat(double lat)
 {
-    return fabsf(lat) <= 90;
+    return fabs(lat) <= 90;
 }
 bool check_lng(double lng)
 {
-    return fabsf(lng) <= 180;
+    return fabs(lng) <= 180;
 }
 bool check_lat(int32_t lat)
 {
